Implemented package update

+++ APIpackages.js
@@ -29,7 +29,11 @@ async function deletePackage(req, res) {
         return;
     }
 
-    const id = req.params.id;
+    const id = req.params?.id;
+    if (id === undefined) {
+        res.status(400).send("ID not defined");
+        return;
+    }
     const deleted = await db.deletePackage(id);
 
     if (!deleted) {
@@ -40,13 +44,16 @@ async function deletePackage(req, res) {
 }
 
 async function getPackage(req, res) {
-    //TODO: GET metadata
     if (!(await checkAuth(req.headers, false))) {
         res.status(401).send();
         return;
     }
 
-    const id = req.params.id;
+    const id = req.params?.id;
+    if (id === undefined) {
+        res.status(400).send("ID not defined");
+        return;
+    }
     const readStream = await db.downloadPackage(id);
 
     if (readStream === null) {
@@ -60,38 +67,51 @@ async function getPackage(req, res) {
     readStream.pipe(res);
 }
 
-async function addPackage(req, res) {
+async function updatePackage(req, res) {
     if (!(await checkAuth(req.headers, false))) {
         res.status(401).send();
         return;
     }
 
-    var metadata = req.body.metadata;
+    const id = req.params?.id;
+    if (id === undefined) {
+        res.status(400).send("no ID provided");
+        return;
+    }
+
+    // if (!(await db.checkPackage(id))) {
+    //     res.status(404).send("package not found");
+    //     return;
+    // }
+
+    var newMetadata = req.body?.metadata;
     const packageUrl = req.body?.data?.URL;
     const content = req.body?.data?.Content;
 
+    if (newMetadata === undefined) {
+        res.status(400).send("no metadata provided");
+        return;
+    }
+
     if (!packageUrl && !content) {
-        res.status(400);
+        res.status(400).send();
+        return;
     }
 
-    metadata.URL = packageUrl;
-    if (packageUrl && !content) {
-        //TODO: get github link out of zip
-        // const canIngest = await checkIngestibility(await rate(packageUrl));
-        // if (!canIngest) {
-        //     res.status(200).send("Could not ingest because bad package");
-        //     return;
-        // }
-
-        metadata = await saveRepo(packageUrl, metadata);
-    }
-
-    if (packageUrl && content) {
-        const contentBuf = Buffer.from(content, "base64"); // Ta-da
-        metadata = await saveZip(contentBuf, metadata);
-    }
-
+    let metadata = await db.getPackageMetadata(id);
     if (!metadata) {
+        res.status(404).send("Package does not exist");
+        return;
+    }
+
+    if (!checkMetadata(metadata, newMetadata)) {
+        res.status(400).send("Metadata does not match");
+        return;
+    }
+
+    let success = upload(packageUrl, content, metadata);
+
+    if (!success) {
         res.status(400).send();
         return;
     }
@@ -100,9 +120,78 @@ async function addPackage(req, res) {
     res.json(metadata);
 }
 
-async function saveZip(contentBuf, metadata) {
-    metadata = db.uploadPackageLocal(contentBuf, metadata);
+function checkMetadata(oldData, newData) {
+    if (!oldData || !newData) {
+        return false;
+    }
+    return (
+        oldData.ID === newData.ID &&
+        oldData.Version === newData.Version &&
+        oldData.Name === newData.Name
+    );
+}
 
+async function addPackage(req, res) {
+    if (!(await checkAuth(req.headers, false))) {
+        res.status(401).send();
+        return;
+    }
+
+    var metadata = req.body?.metadata;
+    const packageUrl = req.body?.data?.URL;
+    const content = req.body?.data?.Content;
+
+    if (metadata === undefined) {
+        res.status(400).send("no metadata provided");
+        return;
+    }
+
+    if (await db.checkPackage(metadata.ID)) {
+        res.status(400).send("package with ID already exists");
+        return;
+    }
+
+    if (!packageUrl && !content) {
+        res.status(400).send();
+        return;
+    }
+
+    metadata.URL = packageUrl;
+
+    let success = upload(packageUrl, content, metadata);
+
+    if (!success) {
+        res.status(400).send();
+        return;
+    }
+    db.savePackageMetadata(metadata);
+
+    res.status(200);
+    res.json(metadata);
+}
+
+async function upload(packageUrl, content, metadata) {
+    let success = false;
+    if (packageUrl && !content) {
+        //TODO: get github link out of zip
+        // const canIngest = await checkIngestibility(await rate(packageUrl));
+        // if (!canIngest) {
+        //     res.status(200).send("Could not ingest because bad package");
+        //     return;
+        // }
+
+        success = await saveRepo(packageUrl, metadata);
+    }
+
+    if (packageUrl && content) {
+        const contentBuf = Buffer.from(content, "base64"); // Ta-da
+        success = await saveZip(contentBuf, metadata);
+    }
+    return success;
+}
+
+async function saveZip(contentBuf, metadata) {
+    metadata = await db.uploadPackageLocal(contentBuf, metadata);
     return metadata;
 }
 
@@ -124,7 +213,7 @@ async function saveRepo(url, metadata) {
     }
 
     const downloadUrl = `https://codeload.github.com/${owner}/${name}/zip/heads/${defaultBranch}`;
-    metadata = db.uploadPackagePublic(downloadUrl, metadata);
+    metadata = await db.uploadPackagePublic(downloadUrl, metadata);
 
     return metadata;
 }
@@ -163,6 +252,7 @@ module.exports = {
     getPackage,
     deletePackage,
     getPackages,
+    updatePackage,
 };
 
 /* *******************************************
